// src/termor_data/schema.ts
export type Usage = {
    owner: string
    status: string
    costs: number
    region: string
    stability: number
    lastEdited: string
};

export type OverviewData = {
    date: string
    'Rows written': number
    'Rows read': number
    Queries: number
    'Payments completed': number
    'Sign ups': number
    Logins: number
};